# ClinicService/Models/Requests/CreatePetRequest.cs
namespace ClinicService.Models.Requests
{
    public class CreatePetRequest
    {
        public int PetId { get; set; }

        public int ClientId { get; set; }

        public string Name { get; set; }

        public DateTime Birthday { get; set; }
    }
}